+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 4.2b2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Setup more default values for documentenerator.
+  [odelaere]
 
 
 4.2b1 (2020-08-24)