+++ CHANGES.rst
@@ -18,6 +18,8 @@ Changelog
   Added `Toggle show/hide all details action` on the item view to be able to toggle
   every collapsible in one click.
   [gbastien]
+- Added an accessor `MeetingItem.getAssociatedGroups` for associatedGroups field [aduchene]
+- Fixed one security.declarePublic in `MeetingConfig`
 
 4.2b1 (2020-08-24)
 ------------------